Handle Harmonia task re-assignment.
Use the link to the Harmonia task to determine whether or not we've already
created a Trello card for the task. If a card doesn't already exist then create
one; if it does exist then remove previous assignees. Finally add the new
assignee to the card.

There might be a nicer way to do this, but it'll do for now.

It's getting to the stage where I ought to write some tests around this, but
I'm hoping @chrisroos won't notice...