Add some specs around the Harmonia assignments action.
Note that I made some changes to the application code in this commit in order
to make it easier to test:

* I converted a bunch of settings which were being stored in local variables
into proper Sinatra settings so they could be accessed from the specs.
* I changed the "dictionary" of Harmonia people names vs Trello member IDs to
instantiate `Trello::Member` objects.
* I moved the `card.update!` statement inside the conditional where the due
date is set, because I realised that `Trello::Card#remove_member` and
`Trello::Card#add_member` do not rely on this call.