Explicitly set the downstream app to nil.
Previously the named argument was being passed to the superclass `#initialize`
method i.e. `{ harmonia: <harmonia-instance> }`.

When a route was not found, the fact that the downstream app was set to a
non-nil value meant that Sinatra attempted to forward the request. However,
since the `Hash` did not respond to `call`, this resulted in the following
exception:

    RuntimeError - downstream app not set

By setting the downstream app to nil, Sinatra does not attempt to forward
the request and raises a `Sinatra::NotFound` exception as intended.