Avoid creating the 'spec' task in production
I'm unable to run `rake` on production because 'rspec' is only
available in the 'test' group. I want to run `rake airbrake:test` on
production and am hoping that this change will allow me to do so. I've
checked that running `rake` in development still executes the tests as
expected.