Respond to Trello webhook request with 200 OK if card not found
If a Harmonia-related card is archived and then deleted immediately afterwards
it's possible for Trello to send a webhook request for the archive change, but
for the card to have been deleted by the time we try to add a comment to it.
I believe this is what happened here [1].

Since marking the Harmonia task as done is the important part of the action, it
seems OK to rescue this Trello exception and respond to Trello with a 200 OK.
This should avoid Trello pointlessly retrying the webhook request (and in the
process causing the Harmonia Slack bot to keep reporting the task being marked
as done).

Note that the exception will still be recorded in Errbit using Airbrake. We
should be able to tell the exception was handled by looking at the line number
in the stack trace.

[1]: https://gofreerange-errbit.herokuapp.com/apps/542ae6297b273bee6c000050/problems/55bba16f1b0e0fc6c5000030