Move cards with earliest due dates nearest the top
The approach I was taking before didn't seem to be working correctly. I assume
this was because when the `pos` value was set for each card, the `pos` values of
other cards might've changed or even the value set might not be the final value.

Anyway, I hope this approach is more robust. Only cards with due dates are now
moved. We then iterate through these cards in reverse chronological order and
move each one to the top of the list. The theory is that the card with the
earliest due date will end up at the top; whereas the card with the latest due
date will end up at the bottom of the cards with due dates.