Handle Harmonia task re-assignment when card not in Harmonia list
Fixes #19.

When handling a web hook callback from Harmonia, the related Trello card
may not be in the list in which it was created (currently the
"Harmonia" list), e.g. if someone has moved it to the "To Do" list.

Previously the app would only look for the card in the "Harmonia" list
and if the card was not found, a new one would be created. Now the app
looks for the card in the board as a whole, i.e. not just in one list.