Stop sorting cards after new card creation
The sorting seems less relevant now that we have a separate Trello list
for Harmonia cards; putting new cards at the bottom of the list should
achieve a pretty similar effect. This commit effectively reverts this
old commit [1], although new cards are added at the bottom to reflect
the current ordering.

Sorting the cards results in quite a few Trello API requests, especially
when there are a lot of cards in the list. My suspicion is that this
might mean we hit the Trello API rate limit [2] from time to time,
possibly resulting in `Trello::Error` exceptions like this one [3].

[1]: 7ac7990bb6a662be6b3345386b572388a03da0e7
[2]: http://help.trello.com/article/838-api-rate-limits
[3]: https://gofreerange-errbit.herokuapp.com/apps/542ae6297b273bee6c000050/problems/592bd5563c352c0008957467